calendar logic and styling
