+++ src/components/Pages/Dashboard/Calendar/Calendar.tsx
@@ -1,82 +1,9 @@
-import {
-  Button,
-  Dialog,
-  DialogActions,
-  DialogContent,
-  DialogContentText,
-  DialogTitle,
-  Typography,
-} from "@mui/material";
+import { Typography } from "@mui/material";
 import { Box } from "@mui/system";
-import { useSelector, useDispatch } from "react-redux";
-import { toggleDone } from "../../../../app/features/habit/habitSlice";
+import { useSelector } from "react-redux";
 import { RootState } from "../../../../app/store";
-import CheckCircleIcon from "@mui/icons-material/CheckCircle";
 import theme from "../../../../styles/theme";
-import { useState } from "react";
-
-const Day = ({
-  done,
-  index,
-  reward,
-}: {
-  done: boolean;
-  index: number;
-  reward?: { label: string; day: number };
-}) => {
-  const dispatch = useDispatch();
-  const [rewardVisible, setRewardVisible] = useState(false);
-  return (
-    <Box
-      sx={{
-        color: theme.palette.common.black,
-        borderRadius: "100%",
-        background: done ? theme.palette.secondary.main : "white",
-        width: "3rem",
-        height: "3rem",
-        display: "flex",
-        alignItems: "center",
-        justifyContent: "center",
-        cursor: "pointer",
-      }}
-      onClick={() => {
-        if (!done && reward) {
-          setRewardVisible(true);
-        }
-        dispatch(toggleDone(index));
-      }}
-    >
-      {done && <CheckCircleIcon />}
-      {!done && <Typography variant="subtitle2">{index + 1}</Typography>}
-      <Dialog
-        open={rewardVisible}
-        onClose={() => setRewardVisible(false)}
-        aria-labelledby="alert-dialog-title"
-        aria-describedby="alert-dialog-description"
-        sx={{ ".MuiPaper-root": { padding: "2rem 4rem" } }}
-      >
-        <DialogTitle id="alert-dialog-title">{"Reward!"}</DialogTitle>
-        <DialogContent>
-          <DialogContentText id="alert-dialog-description">
-            {reward?.label}
-          </DialogContentText>
-        </DialogContent>
-        <DialogActions>
-          <Button
-            onClick={(e) => {
-              e.stopPropagation();
-              setRewardVisible(false);
-            }}
-            variant="contained"
-            color="secondary"
-          >
-            Close
-          </Button>
-        </DialogActions>
-      </Dialog>
-    </Box>
-  );
-};
+import Day from "./CalendarDay/CalendarDay";
 
 const Calendar = () => {
   const habit = useSelector((state: RootState) => state.habit);
@@ -106,6 +33,7 @@ const Calendar = () => {
             key={index}
             index={index}
             reward={habit.rewards.find((el) => el.day === index + 1)}
+            startDate={habit.startDate}
           />
         ))}
       </Box>
